requirements.txt
pillow==9.5.0
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
anthropic>=1.0.0
pdf2image==1.16.3
numpy==1.24.3
python-dotenv==1.0.0
httpx<0.27.0
